Fixed Precision Error in Cameras

+++ src/j2d/engine/camera/Camera.java
@@ -2,6 +2,6 @@ package j2d.engine.camera;
 
 public abstract class Camera {
 
-    abstract int getOffsetX();
-    abstract int getOffsetY();
+    abstract double getOffsetX();
+    abstract double getOffsetY();
 }
